+++ R/check_method_params.R
@@ -36,11 +36,11 @@ check_method_params <- function(y, ite, params) {
     ite_method_dis <- "aipw"
   } else {
     if (!(ite_method_dis %in% c("aipw", "slearner", "tlearner", "xlearner",
-                                "bart", "bcf", "cf", "tpoisson"))) {
+                                "bart", "cf", "tpoisson"))) {
       stop(paste(
         "Invalid ITE method for discovery subsample. Please choose ",
         "from the following:\n", "'aipw', 'bart', 'slearner','tlearner', ",
-        "'xlearner', 'bcf', 'cf', or 'tpoisson'"
+        "'xlearner', 'cf', or 'tpoisson'"
       ))
     }
   }
@@ -51,11 +51,11 @@ check_method_params <- function(y, ite, params) {
     ite_method_inf <- "aipw"
   } else {
     if (!(ite_method_inf %in% c("aipw", "slearner", "tlearner", "xlearner",
-                                "bart", "bcf", "cf", "tpoisson"))) {
+                                "bart", "cf", "tpoisson"))) {
       stop(paste(
         "Invalid ITE method for inference subsample. Please choose ",
         "from the following:\n", "'aipw', 'bart', 'slearner','tlearner', ",
-        "'xlearner', 'bcf', 'cf', or 'tpoisson'"
+        "'xlearner', 'cf', or 'tpoisson'"
       ))
     }
   }